@@ -1,9 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-
-        <IsPackable>false</IsPackable>
+      <TargetFramework>net6.0</TargetFramework>
+      <IsPackable>false</IsPackable>
     </PropertyGroup>
   
   <ItemGroup>
@@ -19,6 +18,10 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
+  </ItemGroup>
+  
+  <ItemGroup>
+    <ProjectReference Include="..\ReactiveMarbles.Extensions\ReactiveMarbles.Extensions.csproj" />
   </ItemGroup>
 
 </Project>
