@@ -5,30 +5,29 @@
 using System;
 using System.Reactive.Disposables;
 
-namespace ReactiveMarbles.Extensions
+namespace ReactiveMarbles.Extensions;
+
+/// <summary>
+/// Extension methods associated with the IDisposable interface.
+/// </summary>
+public static class DisposableExtensions
 {
     /// <summary>
-    /// Extension methods associated with the IDisposable interface.
+    /// Ensures the provided disposable is disposed with the specified <see cref="CompositeDisposable"/>.
     /// </summary>
-    public static class DisposableExtensions
+    /// <typeparam name="T"> The type of the disposable. </typeparam>
+    /// <param name="item"> The disposable we are going to want to be disposed by the CompositeDisposable. </param>
+    /// <param name="compositeDisposable"> The <see cref="CompositeDisposable"/> to which <paramref name="item"/> will be added. </param>
+    /// <returns> The disposable.</returns>
+    public static T DisposeWith<T>(this T item, CompositeDisposable compositeDisposable)
+        where T : IDisposable
     {
-        /// <summary>
-        /// Ensures the provided disposable is disposed with the specified <see cref="CompositeDisposable"/>.
-        /// </summary>
-        /// <typeparam name="T"> The type of the disposable. </typeparam>
-        /// <param name="item"> The disposable we are going to want to be disposed by the CompositeDisposable. </param>
-        /// <param name="compositeDisposable"> The <see cref="CompositeDisposable"/> to which <paramref name="item"/> will be added. </param>
-        /// <returns> The disposable.</returns>
-        public static T DisposeWith<T>(this T item, CompositeDisposable compositeDisposable)
-            where T : IDisposable
+        if (compositeDisposable is null)
         {
-            if (compositeDisposable is null)
-            {
-                throw new ArgumentNullException(nameof(compositeDisposable));
-            }
-
-            compositeDisposable.Add(item);
-            return item;
+            throw new ArgumentNullException(nameof(compositeDisposable));
         }
+
+        compositeDisposable.Add(item);
+        return item;
     }
 }
